# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Synie API
  version: 0.1.0
  description: Synie ERP Go 主 API。金额与小数在 JSON 中一律使用字符串。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /files/query:
    post:
      operationId: querySysFiles
      tags: [files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ListQuery" }
      responses:
        "200":
          description: 文件对象分页结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoredFileList" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /files:
    post:
      operationId: uploadFile
      tags: [files]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/FileUpload" }
      responses:
        "201":
          description: 文件已上传，可选同时挂接业务宿主
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileUploadResult" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /files/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      operationId: downloadFile
      tags: [files]
      responses:
        "200":
          description: 文件字节；对象存储可返回 302 预签名地址
          headers:
            Content-Disposition:
              schema: { type: string }
            X-Content-Type-Options:
              schema: { type: string }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        "302": { description: 对象存储预签名地址 }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteSysFile
      tags: [files]
      responses:
        "204": { description: 文件元数据与对象已删除 }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /files/{id}/metadata:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      operationId: getSysFileMetadata
      tags: [files]
      responses:
        "200":
          description: 文件对象元数据
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoredFile" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /files/{id}/attachments:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      operationId: attachFile
      tags: [files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttachmentCreate" }
      responses:
        "201":
          description: 已补挂附件
          content:
            application/json:
              schema:
                type: object
                required: [attachment]
                properties:
                  attachment: { $ref: "#/components/schemas/Attachment" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /files/attachments/query:
    post:
      operationId: querySysAttachments
      tags: [files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttachmentQuery" }
      responses:
        "200":
          description: 当前 Actor 公司范围内的附件
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttachmentList" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /files/attachments/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    delete:
      operationId: deleteSysAttachment
      tags: [files]
      responses:
        "204": { description: 附件挂接已删除 }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /system/storages/query:
    post:
      operationId: querySysStorages
      tags: [files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ListQuery" }
      responses:
        "200":
          description: 存储接入分页结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorageEndpointList" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /system/storages:
    post:
      operationId: createSysStorage
      tags: [files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StorageEndpointCreate" }
      responses:
        "201":
          description: 存储接入已创建
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorageEndpoint" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /system/storages/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      operationId: getSysStorage
      tags: [files]
      responses:
        "200":
          description: 存储接入（不返回 Secret Access Key）
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorageEndpoint" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateSysStorage
      tags: [files]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StorageEndpointUpdate" }
      responses:
        "200":
          description: 存储接入已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StorageEndpoint" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteSysStorage
      tags: [files]
      responses:
        "204": { description: 存储接入已删除 }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /system/storages/{id}/set-default:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      operationId: setDefaultSysStorage
      tags: [files]
      responses:
        "204": { description: 已设为全局默认存储 }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: 请求失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    StoredFile:
      type: object
      required: [id, storage, key, filename, size, sha256, insertedAt]
      properties:
        id: { type: string, format: uuid }
        storage: { type: string }
        key: { type: string }
        filename: { type: string }
        contentType: { type: string, nullable: true }
        size: { type: integer, format: int64 }
        sha256: { type: string }
        insertedAt: { type: string, format: date-time }
        uploadedById: { type: string, format: uuid, nullable: true }
    StoredFileList:
      type: object
      required: [count, results]
      properties:
        count: { type: integer, format: int64 }
        results:
          type: array
          items: { $ref: "#/components/schemas/StoredFile" }
    FileUpload:
      type: object
      required: [file]
      properties:
        file: { type: string, format: binary }
        ownerType: { type: string }
        ownerId: { type: string, format: uuid }
        category: { type: string }
    FileUploadResult:
      type: object
      required: [file]
      properties:
        file: { $ref: "#/components/schemas/StoredFile" }
        attachment:
          allOf:
            - $ref: "#/components/schemas/Attachment"
          nullable: true
    Attachment:
      type: object
      required: [id, fileId, ownerType, ownerId, category, insertedAt]
      properties:
        id: { type: string, format: uuid }
        fileId: { type: string, format: uuid }
        ownerType: { type: string }
        ownerId: { type: string, format: uuid }
        category: { type: string }
        companyId: { type: string, format: uuid, nullable: true }
        insertedAt: { type: string, format: date-time }
        file:
          allOf:
            - $ref: "#/components/schemas/StoredFile"
          nullable: true
    AttachmentCreate:
      type: object
      required: [ownerType, ownerId]
      properties:
        ownerType: { type: string }
        ownerId: { type: string, format: uuid }
        category: { type: string }
    AttachmentQuery:
      type: object
      properties:
        limit: { type: integer, minimum: 1, maximum: 200, default: 200 }
        offset: { type: integer, minimum: 0, default: 0 }
        fileId: { type: string, format: uuid }
        ownerType: { type: string }
        ownerId: { type: string, format: uuid }
        category: { type: string }
    AttachmentList:
      type: object
      required: [count, results]
      properties:
        count: { type: integer, format: int64 }
        results:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
    StorageKind:
      type: string
      enum: [LOCAL, S3, OSS]
    StorageEndpoint:
      type: object
      required:
        [id, name, label, kind, secretConfigured, builtin, isDefault, insertedAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        label: { type: string }
        kind: { $ref: "#/components/schemas/StorageKind" }
        root: { type: string, nullable: true }
        endpoint: { type: string, nullable: true }
        region: { type: string, nullable: true }
        bucket: { type: string, nullable: true }
        prefix: { type: string, nullable: true }
        accessKeyId: { type: string, nullable: true }
        secretConfigured: { type: boolean }
        builtin: { type: boolean }
        isDefault: { type: boolean }
        insertedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    StorageEndpointList:
      type: object
      required: [count, results]
      properties:
        count: { type: integer, format: int64 }
        results:
          type: array
          items: { $ref: "#/components/schemas/StorageEndpoint" }
    StorageEndpointCreate:
      type: object
      required: [name, label, kind]
      properties:
        name: { type: string, maxLength: 32 }
        label: { type: string, maxLength: 64 }
        kind: { $ref: "#/components/schemas/StorageKind" }
        root: { type: string, nullable: true }
        endpoint: { type: string, nullable: true }
        region: { type: string, nullable: true }
        bucket: { type: string, nullable: true }
        prefix: { type: string, nullable: true }
        accessKeyId: { type: string, nullable: true }
        secretAccessKey: { type: string, writeOnly: true }
    StorageEndpointUpdate:
      type: object
      properties:
        label: { type: string, maxLength: 64 }
        root: { type: string, nullable: true }
        endpoint: { type: string, nullable: true }
        region: { type: string, nullable: true }
        bucket: { type: string, nullable: true }
        prefix: { type: string, nullable: true }
        accessKeyId: { type: string, nullable: true }
        secretAccessKey: { type: string, writeOnly: true }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [unauthorized, rate_limited, forbidden, validation, not_found, conflict, not_implemented, internal]
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    ListQuery:
      type: object
      additionalProperties: false
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 200
          default: 20
        offset:
          type: integer
          minimum: 0
          default: 0
        search:
          type: string
          maxLength: 256
        sort:
          $ref: "#/components/schemas/Sort"
        filter:
          $ref: "#/components/schemas/FilterState"
    Sort:
      type: object
      additionalProperties: false
      required: [column, direction]
      properties:
        column:
          type: string
        direction:
          type: string
          enum: [ascending, descending]
    FilterState:
      type: object
      additionalProperties:
        type: object
        additionalProperties: true
        required: [kind]
        properties:
          kind:
            type: string
            enum: [text, bool, enum, enumArray, number, date, fk, polyFk]
...